[HAWKULAR-732] initial commit for new Bus class including integration tests
[HAWKULAR-732] add support for input streams

[HAWKULAR-732] move files to correct location in src tree

These files were migrated from the hawkular-bus repo.

[HAWKULAR-732] add overloaded versions of methods to take timeout param

[HAWKULAR-732] add support for filtering with message selectors

[HAWKMULAR-732] initial support for listeners including integration test

I had to comment out some code in MessageProcessor and in the BusTest class in
the itest module due to changes in BasicMessageListener.

The Bus.register method explicitly creates a new JMSContext in order to avoid
warnings from Artemis about concurrent access to the JMS session. I was seeing
this in the integration tests, although I am not clear on where/when the
concurrent access is happening.

I have also added a @Predestroy method to make sure any context/consumer
objects that have been created manually get closed.

[HAWKULAR-732] use non-pooled connection factory for message listeners

The JEE spec forbids the use of JMS message listeners (don't ask me why). In
WildFly 10 the message subsystem is implemented on top of Artemis. When you try
to use a message listener in WildFly 10 with a pooled connection factory, it
results in the following exception,

javax.jms.IllegalStateRuntimeException: This method is not applicable inside the application server. See the J2EE spec, e.g. J2EE1.4 Section 6.6
	at org.apache.activemq.artemis.ra.ActiveMQRASession.checkStrict(ActiveMQRASession.java:1452)
	at org.apache.activemq.artemis.ra.ActiveMQRAMessageConsumer.setMessageListener(ActiveMQRAMessageConsumer.java:123)
	at org.apache.activemq.artemis.jms.client.ActiveMQJMSConsumer.setMessageListener(ActiveMQJMSConsumer.java:59)

The solution is to use a non-pooled connection factory for those situations in
which a message listener is needed.

This commit also includes an initial MDB test.

[HAWKULAR-732] remove obsolete classes

This commit removes a number of files that were specific to the former
ActiveMQ implementation. It also removes several unused classes.

Of more significace is the change to BasicMessageListener. The following method
has been removed,

    void onBasicMessage(BasicMessageWithExtraData<T> msgWithExtraData)

and I have added a binaryData property to BasicMessage. It seemed awkward to me
to deal with one message class if your message involves streaming and another
class when no streaming is involved and neither of these classes share a common
base class/interface. Having a binaryData property simplifies the
BasicMessageListener interface some what.

[HAKWULAR-732] Put JSON serialization in MessageSerializer for reuse

In the original impl, AbstractMessage had some static methods that handled
mapping to/from JSON. A new OjbectMapper was created each time we serialized or
deserialized. ObjectMapper creation is (according to the Jackson docs) some
what expensive and should be minimized. MessageSerializer is @ApplictionScoped
which means an ObjectMapper is created only once. It is injected into MDBs that
extend BasicMessageListener. For instances of BasicMessageListener that are not
MDBs, Bus will inject MessageSerializer.

I updated the mapper logic a bit to remain consistent with the prior
implementation. We have at least one place that uses the fail on unknown
properties Jackson feature. I updated MessageSerializer to keep that
behavior.

[HAWKUALR-732] add back support for extra headers and return message id

MessageProcessor provided an API to provide an extra set of headers that would
override headers supplied by the message. I have added that functionality to
Bus.

The send methods in MessageProcessor returned a MessageId. I have updated the
send methods in Bus to also return the MessageId.

[HAWKULAR-732] add support and example test for alternative object mapper

[HAWKULAR-732] clean up message mapping

This commit cleans up the mapping of messages to/from JMS. For sending messages
Bus needs to map properties from BasicMessage to the JMS message. And for
receiving messages, MessageSerializer maps properties from the JMS message to
the BasicMessage. The mapping logic is in MessageSerializer because it is used
both by Bus and by BasicMessageListener when BasicMessageListener is an MDB.

[HAWKULAR-732] adding javadocs

[HAWKULAR-732] removing obsolete classes

Adding new, qualifier annotations in JMSQueue and JMSTopic

JsonMapper contains parsing logic without JMS dependencies. This code is
being using in cmdgw branch, and cmdgw should not be pulling in any JMS
dependencies since it is also used in the agent.